+++ libclamav/others.c
@@ -567,7 +567,7 @@ int cli_checklimits(const char *who, cli_ctx *ctx, unsigned long need1, unsigned
         /* if the remaining scansize is too small... */
         if(ctx->engine->maxscansize-ctx->scansize<needed) {
 	    /* ... we tell the caller to skip this file */
-	    cli_dbgmsg("%s: scansize exceeded (initial: %lu, remaining: %lu, needed: %lu)\n", who, ctx->engine->maxscansize, ctx->scansize, needed);
+	    cli_dbgmsg("%s: scansize exceeded (initial: %lu, remaining: %lu, needed: %lu)\n", who, (unsigned long int) ctx->engine->maxscansize, (unsigned long int) ctx->scansize, needed);
 	    ret = CL_EMAXSIZE;
 	}
     }
@@ -575,7 +575,7 @@ int cli_checklimits(const char *who, cli_ctx *ctx, unsigned long need1, unsigned
     /* if we have per-file size limits, and we are overlimit... */
     if(needed && ctx->engine->maxfilesize && ctx->engine->maxfilesize<needed) {
 	/* ... we tell the caller to skip this file */
-        cli_dbgmsg("%s: filesize exceeded (allowed: %lu, needed: %lu)\n", who, ctx->engine->maxfilesize, needed);
+        cli_dbgmsg("%s: filesize exceeded (allowed: %lu, needed: %lu)\n", who, (unsigned long int) ctx->engine->maxfilesize, needed);
 	ret = CL_EMAXSIZE;
     }
